requirements.txt
cert-core>=2.1.10
bloxberg-cert-schema
configargparse>=0.13.0
jsonpath-rw>=1.4.0
pycoin>=0.80
tox>=3.0.0
httpx
aiofiles
PyMuPDF
pyqrcode
pypng
lds-merkle-proof-2019-bloxberg
pytest
